[Refactor] 기본 엔진(순환, 랜덤 편도, 최단 경로)만을 유지
