+++ src/route_engine/engines/__init__.py
@@ -1,7 +1,3 @@
-from src.route_engine.engines.circular.child   import CircularChildEngine
-from src.route_engine.engines.circular.random  import CircularRandomEngine
-from src.route_engine.engines.circular.running import CircularRunningEngine
-from src.route_engine.engines.oneway.child     import OnewayChildEngine
-from src.route_engine.engines.oneway.dijkstra  import OnewayDijkstraEngine
-from src.route_engine.engines.oneway.random    import OnewayRandomEngine
-from src.route_engine.engines.oneway.running   import OnewayRunningEngine
+from src.route_engine.engines.circular_random  import CircularRandomEngine
+from src.route_engine.engines.dijkstra  import OnewayDijkstraEngine
+from src.route_engine.engines.oneway_random    import OnewayRandomEngine
