+++ src/route_engine/engines/path_utils.py
@@ -132,130 +132,3 @@ class PathUtils:
             (self.G.get_edge_data(nodes[i], nodes[i + 1]) or {}).get("length", 0)
             for i in range(len(nodes) - 1)
         )  # 인접 노드 쌍의 length 합산
-
-    def circular_random_walk(self, start_node: int, target_km: float = 3.0) -> list[int]:
-        """
-        custom_score 기반 확률적 랜덤 워크로 순환 경로 노드 목록을 반환합니다.
-        """
-        target_m   = target_km * 1000  # 목표 거리(미터)
-        path_nodes = [start_node]
-        total_dist = 0.0
-        current    = start_node
-        visited: dict = {}             # 엣지별 방문 횟수
-        sx = self.G.nodes[start_node].get("lon", 0)  # 출발점 경도
-        sy = self.G.nodes[start_node].get("lat", 0)  # 출발점 위도
-
-        while total_dist < target_m * 0.75:  # 목표의 75%까지 탐색
-            neighbors = list(self.G.neighbors(current))
-            if not neighbors:
-                break
-
-            probs = []
-            for n in neighbors:
-                ek         = tuple(sorted([current, n]))               # 무방향 엣지 키
-                visit_pen  = 1.0 / (1 + visited.get(ek, 0) * 7)        # 재방문 페널티
-                degree_pen = 1.0 / (1 + max(0, 3 - self.G.degree(n)))  # 막힌 노드 억제
-                score      = (self.G.get_edge_data(current, n) or {}).get("custom_score", 1.0)
-
-                if total_dist / target_m < 0.7:
-                    # 전반부: 출발점에서 멀수록 score 작아짐 → 확률 높아짐 (탐색 확장)
-                    dx    = self.G.nodes[n].get("lon", 0) - sx
-                    dy    = self.G.nodes[n].get("lat", 0) - sy
-                    dist  = (dx ** 2 + dy ** 2) ** 0.5
-                    score = score / ((dist + 1e-6) ** 2)
-                else:
-                    # 후반부: 출발점에서 멀수록 score 커짐 → 확률 낮아짐 (귀환 유도)
-                    # Dijkstra 복귀 전 워커가 자연스럽게 출발점 방향으로 수렴하도록 함
-                    dx    = self.G.nodes[n].get("lon", 0) - sx
-                    dy    = self.G.nodes[n].get("lat", 0) - sy
-                    dist  = (dx ** 2 + dy ** 2) ** 0.5
-                    score = score * (1.0 + dist ** 2 * 1e6)
-
-                probs.append((1.0 / (score + 1e-6)) * visit_pen * degree_pen)
-
-            total_p = sum(probs)
-            if total_p == 0:
-                break
-
-            next_node   = random.choices(neighbors, weights=[p / total_p for p in probs], k=1)[0]
-            ek          = tuple(sorted([current, next_node]))
-            visited[ek] = visited.get(ek, 0) + 1                                          # 방문 횟수 누적
-            total_dist += (self.G.get_edge_data(current, next_node) or {}).get("length", 0)
-            path_nodes.append(next_node)
-            current = next_node
-
-        if path_nodes[-1] != start_node:  # 출발 노드 미복귀 시 최단 경로로 복귀
-            try:
-                def _return_w(u, v, d):
-                    ek = tuple(sorted([u, v]))
-                    return d.get("length", 1.0) * (1 + visited.get(ek, 0) * 10)  # 기방문 엣지 패널티
-
-                return_path = nx.shortest_path(self.G, path_nodes[-1], start_node, weight=_return_w)
-                path_nodes += return_path[1:]  # 복귀 경로 연결 (중복 노드 제거)
-            except nx.NetworkXNoPath:
-                pass
-
-        return path_nodes
-
-    def oneway_waypoint_path(self, start: int, end: int, target_km: float = 3.0) -> list[int]:
-        """
-        경유 노드를 활용한 우회 편도 경로 노드 목록을 반환합니다.
-        """
-        target_m = target_km * 1000  # 목표 거리(미터)
-        p1       = self.G.nodes[start]
-        p2       = self.G.nodes[end]
-
-        lon1, lat1 = p1.get("lon", 0), p1.get("lat", 0)
-        lon2, lat2 = p2.get("lon", 0), p2.get("lat", 0)
-        dx, dy     = lon2 - lon1, lat2 - lat1
-        dist_se    = math.sqrt(dx ** 2 + dy ** 2)
-
-        # 출발-도착 직선에 수직인 방향으로 중점을 offset하여 실제 우회 루프 유도
-        # 기존 방식(직선거리 비율 필터만)은 직선 근처 노드가 선택되어 우회 효과 약함
-        offset_deg = (target_km * 0.35) / 111.0
-        if dist_se > 1e-9:
-            side = random.choice([1, -1])
-            px   = -dy / dist_se * side  # 수직 단위벡터 (90도 회전)
-            py   =  dx / dist_se * side
-        else:
-            angle = random.uniform(0, 2 * math.pi)
-            px, py = math.cos(angle), math.sin(angle)
-
-        target_lon = (lon1 + lon2) / 2 + px * offset_deg
-        target_lat = (lat1 + lat2) / 2 + py * offset_deg
-
-        candidates = []
-        for node, data in self.G.nodes(data=True):
-            if node in (start, end):
-                continue
-            nlon, nlat = data.get("lon", 0), data.get("lat", 0)
-            d1    = math.sqrt((nlon - lon1) ** 2 + (nlat - lat1) ** 2) * 111000
-            d2    = math.sqrt((nlon - lon2) ** 2 + (nlat - lat2) ** 2) * 111000
-            total = d1 + d2
-            if target_m * 0.6 <= total <= target_m * 0.9:
-                d_target = math.sqrt((nlon - target_lon) ** 2 + (nlat - target_lat) ** 2)
-                candidates.append((node, d_target))
-
-        if not candidates:
-            waypoint = self.find_nearest_node(target_lat, target_lon)
-        else:
-            candidates.sort(key=lambda x: x[1])
-            waypoint = random.choice(candidates[:5])[0]  # 상위 5개 중 랜덤 선택
-
-        try:
-            path1       = nx.shortest_path(self.G, start, waypoint, weight="custom_score")
-            path1_edges = set(zip(path1[:-1], path1[1:]))
-
-            # G 엣지를 직접 수정하지 않고 클로저로 페널티 적용 → 그래프 오염 방지
-            def penalized_weight(u, v, data):
-                base = data.get("custom_score", 1.0)
-                return base * 100.0 if (u, v) in path1_edges or (v, u) in path1_edges else base
-
-            path2 = nx.shortest_path(self.G, waypoint, end, weight=penalized_weight)
-            return path1[:-1] + path2
-
-        except Exception:
-            try:
-                return nx.shortest_path(self.G, start, end, weight="custom_score")  # 경유 실패 시 직선 최단 경로
-            except Exception:
-                return []
